Refactor prompt generation into separate PromptGenerator class
